split op/backup tests into two files
A quick refactor to clean up the code and split
op/backup testing between unit and integration
test sets, so that we don't bloat a single file.
